Extract an AsmSecretStoreTester, for reuse
